requirements.txt
numpy==1.21.4
opencv_python_headless==4.4.0.46
Pillow==8.4.0
streamlit==1.1.0
tensorflow==2.5.0
